+++ odoo/songs/update/update_10_0_15.py
@@ -14,6 +14,19 @@ def create_product_categories(ctx):
 
 
 @anthem.log
+def update_sale_settings(ctx):
+    """Field 'Tax Display' shoud be B2C"""
+    sale_settings = ctx.env["sale.config.settings"]
+    # already present in songs/install/sale.py BSMTS-217
+    vals = {
+        'sale_show_tax': 'total'
+    }
+    setting = sale_settings.create(vals)
+    setting.execute()
+
+
+@anthem.log
 def main(ctx):
     """create categories"""
     create_product_categories(ctx)
+    update_sale_settings(ctx)